move sysdb getAllEvents and getAllNotifications to DAOs
