+++ transact/src/main/java/dev/dbos/transact/database/SystemDatabase.java
@@ -4,7 +4,6 @@ import dev.dbos.transact.Constants;
 import dev.dbos.transact.config.DBOSConfig;
 import dev.dbos.transact.exceptions.*;
 import dev.dbos.transact.json.DBOSSerializer;
-import dev.dbos.transact.json.SerializationUtil;
 import dev.dbos.transact.workflow.ExportedWorkflow;
 import dev.dbos.transact.workflow.ForkOptions;
 import dev.dbos.transact.workflow.GetWorkflowAggregatesInput;
@@ -352,8 +351,7 @@ public class SystemDatabase implements AutoCloseable {
 
   public void recordChildWorkflow(
       String parentId,
-      String childId, // workflowId of the
-      // child
+      String childId, // workflowId of the child
       int functionId, // func id in the parent
       String functionName,
       long startTime) {
@@ -527,56 +525,11 @@ public class SystemDatabase implements AutoCloseable {
   }
 
   public Map<String, Object> getAllEvents(String workflowId) {
-    return dbRetry(
-        () -> {
-          try (var conn = dataSource.getConnection()) {
-            var events = workflowDAO.listWorkflowEvents(conn, workflowId);
-            var result = new LinkedHashMap<String, Object>();
-            for (var event : events) {
-              result.put(
-                  event.key(),
-                  SerializationUtil.deserializeValue(
-                      event.value(), event.serialization(), this.serializer));
-            }
-            return result;
-          }
-        });
+    return dbRetry(() -> workflowDAO.getAllEvents(workflowId));
   }
 
   public List<NotificationInfo> getAllNotifications(String workflowId) {
-    return dbRetry(
-        () -> {
-          var sql =
-              """
-              SELECT topic, message, serialization, created_at_epoch_ms, consumed
-              FROM "%s".notifications
-              WHERE destination_uuid = ?
-              ORDER BY created_at_epoch_ms
-              """
-                  .formatted(this.schema);
-
-          var notifications = new ArrayList<NotificationInfo>();
-          try (var conn = dataSource.getConnection();
-              var stmt = conn.prepareStatement(sql)) {
-            stmt.setString(1, workflowId);
-            try (var rs = stmt.executeQuery()) {
-              while (rs.next()) {
-                var rawTopic = rs.getString("topic");
-                var topic = Constants.DBOS_NULL_TOPIC.equals(rawTopic) ? null : rawTopic;
-                var serialization = rs.getString("serialization");
-                var message =
-                    SerializationUtil.deserializeValue(
-                        rs.getString("message"), serialization, this.serializer);
-                var createdAtEpochMs = rs.getLong("created_at_epoch_ms");
-                var consumed = rs.getBoolean("consumed");
-                notifications.add(
-                    new NotificationInfo(
-                        topic, message, SystemDatabase.toInstant(createdAtEpochMs), consumed));
-              }
-            }
-          }
-          return notifications;
-        });
+    return dbRetry(() -> notificationsDAO.getAllNotifications(workflowId));
   }
 
   public List<ExportedWorkflow> exportWorkflow(String workflowId, boolean exportChildren) {
